Wire testkit conformance onto javaclient backend + fix bugs uncovered
Adds JavaMetricRegistryTestkitSuite which extends the testkit's cross-backend
MetricRegistrySuite[PrometheusRegistry] + CallbackRegistrySuite[PrometheusRegistry].
Wiring the testkit uncovered two real backend bugs and one v5 API shape that
doesn't fit v6.

Backend bugs fixed:
  - configureBuilderOrRetrieve didn't check callbackState before metric
    registration, so cross-axis (metric vs callback) name collisions were
    silently allowed. Mirrors the v5 check.
  - registerCallback keyed StateKey under name.show (full counter name,
    e.g. foo_total) while metric registration keyed under counterName(name)
    (stripped, foo). Same-name collision detection missed counter cases.
    Aligned to the stripped form.
  - Error messages also surfaced the stripped form to users; added a
    renderedName parameter so error text uses the wire-level name
    (foo_total) like v5.

API shape change (Summary.Value.count: Double -> Long):
  - Upstream prometheus-metrics-core 1.x SummaryDataPointSnapshot constructor
    takes long count; v5 simpleclient took double. Summary.Value previously
    typed count as the value-type parameter A (Double or Long); v6 fixes
    count to Long for both Value[Double] and Value[Long]. Counts are integer
    in Prometheus. Source-incompatible at every Summary.Value(...) construction
    site (compile errors, no silent breaks).

Other v6 behaviour alignments:
  - createAndRegisterInfo registered under name.value (no prefix); snapshot
    lookups via NameUtils.makeName(prefix, name) failed to find it.
    Corrected to register under the prefix-qualified name.
  - registerDoubleHistogramCallback / metricCollectionToSnapshots: convert
    cumulative bucket counts to per-bucket counts (upstream's
    ClassicHistogramBuckets.of expects per-bucket; we were passing
    cumulative).

Five name-collision contract tests promoted from javasimpleclient into the
new javaclient JVM-specific suite (kept here, not in the cross-backend testkit,
because they need a single registry instance implementing both MetricRegistry
and CallbackRegistry; the testkit's separate resources would yield disjoint
internal state and defeat the cross-axis collision check).

Squashes the WIP testkit-wiring checkpoint commit (ba57b46) into this fix.

Verified: 102 of 104 tests pass; the 2 remaining failures are in the legacy
javasimpleclient suite (Long counts > 2^53 lose precision through simpleclient's
Double-typed SummaryMetricFamily wire format). Those resolve when the legacy
backend is removed in the next commit.

+++ modules/prometheus4cats-java/src/test/scala/prometheus4cats/javasimpleclient/JavaMetricRegistrySuite.scala
@@ -186,7 +186,7 @@ class JavaMetricRegistrySuite
       help: Metric.Help,
       commonLabels: CommonLabels,
       extraLabels: Map[Label.Name, String]
-  ): IO[(Option[Map[String, Double]], Option[Double], Option[Double])] = IO {
+  ): IO[(Option[Map[String, Double]], Option[Long], Option[Double])] = IO {
     val n = NameUtils.makeName(prefix, name)
 
     val allLabels = (commonLabels.value ++ extraLabels).map { case (n, v) => n.value -> v }
@@ -212,7 +212,7 @@ class JavaMetricRegistrySuite
 
     (
       quantiles,
-      getMetricValue(state, prefix, show"${name}_count", commonLabels, extraLabels).map(_._1),
+      getMetricValue(state, prefix, show"${name}_count", commonLabels, extraLabels).map(_._1.toLong),
       getMetricValue(state, prefix, show"${name}_sum", commonLabels, extraLabels).map(_._1)
     )
   }
